fix(deps): eliminate RUSTSEC-2026-0258 (h2 0.3) by migrating igd->igd-next and hyper 0.14->1.x
- citadel_wire: replace igd 0.12 (aio) with igd-next 0.17 (aio_tokio) in
  upnp_handler; Gateway is now Gateway<Tokio>, gateway.addr is a SocketAddr,
  and get_external_ip returns IpAddr upstream (still surfaced as Ipv4Addr,
  rejecting IPv6 externals like before)
- citadel_sdk: hyper 0.14 dev-dep -> hyper 1 + hyper-util + http-body-util;
  internal_service tests/doc examples now use hyper::server::conn::http1,
  hyper::client::conn::http1::handshake, TokioIo, Full<Bytes>/Incoming and
  BodyExt::collect instead of Http/client::conn::Builder/Body/to_bytes
- split internal_service.rs (368 lines) into internal_service/{mod.rs,
  tests/mod.rs, tests/http.rs} (all <=250 lines), exact piecewise copy
  aside from the hyper 1.x API swap

Both h2 0.3 paths (igd->hyper 0.14 and the hyper 0.14 workspace dep) are
gone; only h2 0.4.18 (>=0.4.16, patched) remains in the resolved graph.
cargo deny check advisories passes.
